ppi test

+++ PPI/deepinfomax_gae.py
@@ -201,7 +201,7 @@ class GcnInfomax(nn.Module):
                 x, edge_index, batch = data.x, data.edge_index, data.batch
 
 
-                node_mu, node_logvar = self.encoder(x, edge_index, batch)
+                node_mu, node_logvar = self.encoder(x.double(), edge_index, batch)
 
 
                 node_latent_embeddings = reparameterize(training=False, mu=node_mu, logvar=node_logvar)
@@ -369,7 +369,7 @@ if __name__ == '__main__':
 
 
                 optimizer.zero_grad()
-                recon_loss, kl_class, kl_node = model(data.x, data.edge_index, data.batch, data.num_graphs)
+                recon_loss, kl_class, kl_node = model(data.x.double(), data.edge_index, data.batch, data.num_graphs)
                 recon_loss_all += recon_loss
                 kl_class_loss_all += kl_class
                 kl_node_loss_all += kl_node
